fix: file structure renaming and added sliders
