+++ app/src/main/AndroidManifest.xml
@@ -23,7 +23,7 @@
         android:theme="@style/AppTheme">
 
         <activity
-            android:name="app.mlauncher.MainActivity"
+            android:name="com.github.hecodes2much.mlauncher.MainActivity"
             android:configChanges="uiMode"
             android:excludeFromRecents="true"
             android:launchMode="singleTask"
@@ -40,7 +40,7 @@
         </activity>
 
         <activity
-            android:name="app.mlauncher.helper.FakeHomeActivity"
+            android:name="com.github.hecodes2much.mlauncher.helper.FakeHomeActivity"
             android:enabled="false"
             android:launchMode="singleTask"
             android:windowSoftInputMode="stateHidden"
@@ -55,7 +55,7 @@
         </activity>
 
         <receiver
-            android:name="app.mlauncher.listener.DeviceAdmin"
+            android:name="com.github.hecodes2much.mlauncher.listener.DeviceAdmin"
             android:permission="android.permission.BIND_DEVICE_ADMIN"
             android:exported="false">
             <meta-data
@@ -67,7 +67,7 @@
         </receiver>
 
         <service
-            android:name="app.mlauncher.helper.ActionService"
+            android:name="com.github.hecodes2much.mlauncher.helper.ActionService"
             android:label="@string/accessibility_service_name"
             android:permission="android.permission.BIND_ACCESSIBILITY_SERVICE"
             android:exported="false">
