--- app/src/main/java/app/mlauncher/helper/Utils.kt
@@ -1,413 +1,0 @@
-package app.mlauncher.helper
-
-import android.app.Activity
-import android.content.*
-import android.content.pm.LauncherApps
-import android.content.pm.PackageManager
-import android.content.pm.PackageManager.NameNotFoundException
-import android.content.res.Configuration
-import android.content.res.Resources
-import android.net.Uri
-import android.os.Build
-import android.os.Bundle
-import android.os.UserHandle
-import android.os.UserManager
-import android.provider.AlarmClock
-import android.provider.CalendarContract
-import android.provider.MediaStore
-import android.provider.Settings
-import android.util.DisplayMetrics
-import android.util.Log.*
-import android.util.TypedValue
-import android.view.Gravity
-import android.view.View
-import android.view.WindowInsets
-import android.view.WindowManager
-import android.widget.Toast
-import androidx.core.app.ActivityCompat
-import app.mlauncher.BuildConfig
-import app.mlauncher.data.AppModel
-import app.mlauncher.data.Constants.BACKUP_READ
-import app.mlauncher.data.Constants.BACKUP_WRITE
-import app.mlauncher.data.Prefs
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
-import java.text.Collator
-import java.util.*
-import kotlin.math.pow
-import kotlin.math.sqrt
-
-fun showToastLong(context: Context, message: String) {
-    val toast = Toast.makeText(context.applicationContext, message, Toast.LENGTH_LONG)
-    toast.setGravity(Gravity.CENTER, 0, 0)
-    toast.show()
-}
-
-fun showToastShort(context: Context, message: String) {
-    val toast = Toast.makeText(context.applicationContext, message, Toast.LENGTH_SHORT)
-    toast.setGravity(Gravity.CENTER, 0, 0)
-    toast.show()
-}
-
-suspend fun getAppsList(context: Context, showHiddenApps: Boolean = false): MutableList<AppModel> {
-    return withContext(Dispatchers.IO) {
-        val appList: MutableList<AppModel> = mutableListOf()
-
-        try {
-            if (!Prefs(context).hiddenAppsUpdated) upgradeHiddenApps(Prefs(context))
-            val hiddenApps = Prefs(context).hiddenApps
-
-            val userManager = context.getSystemService(Context.USER_SERVICE) as UserManager
-            val launcherApps = context.getSystemService(Context.LAUNCHER_APPS_SERVICE) as LauncherApps
-            val collator = Collator.getInstance()
-
-            val prefs = Prefs(context)
-
-            for (profile in userManager.userProfiles) {
-                for (app in launcherApps.getActivityList(null, profile)) {
-
-
-                    // we have changed the alias identifier from app.label to app.applicationInfo.packageName
-                    // therefore, we check if the old one is set if the new one is empty
-                    val appAlias = prefs.getAppAlias(app.applicationInfo.packageName).ifEmpty {
-                        prefs.getAppAlias(app.label.toString())
-                    }
-
-                    if (showHiddenApps && app.applicationInfo.packageName != BuildConfig.APPLICATION_ID) {
-                        val appModel = AppModel(
-                            app.label.toString(),
-                            collator.getCollationKey(app.label.toString()),
-                            app.applicationInfo.packageName,
-                            app.componentName.className,
-                            profile,
-                            appAlias,
-                        )
-                        appList.add(appModel)
-                    } else if (!hiddenApps.contains(app.applicationInfo.packageName + "|" + profile.toString())
-                        && app.applicationInfo.packageName != BuildConfig.APPLICATION_ID
-                    ) {
-                        val appModel = AppModel(
-                            app.label.toString(),
-                            collator.getCollationKey(app.label.toString()),
-                            app.applicationInfo.packageName,
-                            app.componentName.className,
-                            profile,
-                            appAlias,
-                        )
-                        appList.add(appModel)
-                    }
-
-                }
-            }
-
-            appList.sortBy {
-                if (it.appAlias.isEmpty()) it.appLabel.lowercase() else it.appAlias.lowercase()
-            }
-
-        } catch (e: java.lang.Exception) {
-            d("appList", e.toString())
-        }
-        appList
-    }
-}
-
-suspend fun getHiddenAppsList(context: Context): MutableList<AppModel> {
-    return withContext(Dispatchers.IO) {
-        val pm = context.packageManager
-        if (!Prefs(context).hiddenAppsUpdated) upgradeHiddenApps(Prefs(context))
-
-        val hiddenAppsSet = Prefs(context).hiddenApps
-        val appList = mutableListOf<AppModel>()
-        if (hiddenAppsSet.isEmpty()) {
-            return@withContext appList
-        }
-
-        val userManager = context.getSystemService(Context.USER_SERVICE) as UserManager
-        val collator = Collator.getInstance()
-        for (hiddenPackage in hiddenAppsSet) {
-            val appPackage = hiddenPackage.split("|")[0]
-            val userString = hiddenPackage.split("|")[1]
-            var userHandle = android.os.Process.myUserHandle()
-            for (user in userManager.userProfiles) {
-                if (user.toString() == userString) userHandle = user
-            }
-            try {
-                @Suppress("DEPRECATION")
-                val appInfo = pm.getApplicationInfo(appPackage, 0)
-                val appName = pm.getApplicationLabel(appInfo).toString()
-                val appKey = collator.getCollationKey(appName)
-                appList.add(AppModel(appName, appKey, appPackage, "", userHandle, Prefs(context).getAppAlias(appName)))
-            } catch (e: NameNotFoundException) {
-                d("getHiddenAppsList", e.toString())
-            }
-        }
-        appList.sort()
-        appList
-    }
-}
-
-// This is to ensure backward compatibility with older app versions
-// which did not support multiple user profiles
-private fun upgradeHiddenApps(prefs: Prefs) {
-    val hiddenAppsSet = prefs.hiddenApps
-    val newHiddenAppsSet = mutableSetOf<String>()
-    for (hiddenPackage in hiddenAppsSet) {
-        if (hiddenPackage.contains("|")) newHiddenAppsSet.add(hiddenPackage)
-        else newHiddenAppsSet.add(hiddenPackage + android.os.Process.myUserHandle().toString())
-    }
-    prefs.hiddenApps = newHiddenAppsSet
-    prefs.hiddenAppsUpdated = true
-}
-
-fun getUserHandleFromString(context: Context, userHandleString: String): UserHandle {
-    val userManager = context.getSystemService(Context.USER_SERVICE) as UserManager
-    for (userHandle in userManager.userProfiles) {
-        if (userHandle.toString() == userHandleString) {
-            return userHandle
-        }
-    }
-    return android.os.Process.myUserHandle()
-}
-
-fun ismlauncherDefault(context: Context): Boolean {
-    val launcherPackageName = getDefaultLauncherPackage(context)
-    return BuildConfig.APPLICATION_ID == launcherPackageName
-}
-
-fun getDefaultLauncherPackage(context: Context): String {
-    val intent = Intent()
-    intent.action = Intent.ACTION_MAIN
-    intent.addCategory(Intent.CATEGORY_HOME)
-    val packageManager = context.packageManager
-    @Suppress("DEPRECATION")
-    val result = packageManager.resolveActivity(intent, 0)
-    return if (result?.activityInfo != null) {
-        result.activityInfo.packageName
-    } else "android"
-}
-
-// Source: https://stackoverflow.com/a/13239706
-fun resetDefaultLauncher(context: Context) {
-    try {
-        val packageManager = context.packageManager
-        val componentName = ComponentName(context, FakeHomeActivity::class.java)
-        packageManager.setComponentEnabledSetting(
-            componentName,
-            PackageManager.COMPONENT_ENABLED_STATE_ENABLED,
-            PackageManager.DONT_KILL_APP
-        )
-        val selector = Intent(Intent.ACTION_MAIN)
-        selector.addCategory(Intent.CATEGORY_HOME)
-        context.startActivity(selector)
-        packageManager.setComponentEnabledSetting(
-            componentName,
-            PackageManager.COMPONENT_ENABLED_STATE_DISABLED,
-            PackageManager.DONT_KILL_APP
-        )
-    } catch (e: Exception) {
-        e.printStackTrace()
-    }
-}
-
-fun openAppInfo(context: Context, userHandle: UserHandle, packageName: String) {
-    val launcher = context.getSystemService(Context.LAUNCHER_APPS_SERVICE) as LauncherApps
-    val intent: Intent? = context.packageManager.getLaunchIntentForPackage(packageName)
-    intent?.let {
-        launcher.startAppDetailsActivity(intent.component, userHandle, null, null)
-    } ?: showToastShort(context, "Unable to to open app info")
-}
-
-fun openDialerApp(context: Context) {
-    try {
-        val sendIntent = Intent(Intent.ACTION_DIAL)
-        context.startActivity(sendIntent)
-    } catch (e: java.lang.Exception) {
-        d("openDialerApp", e.toString())
-    }
-}
-
-fun openCameraApp(context: Context) {
-    try {
-        val sendIntent = Intent(MediaStore.INTENT_ACTION_STILL_IMAGE_CAMERA)
-        context.startActivity(sendIntent)
-    } catch (e: java.lang.Exception) {
-        d("openCameraApp", e.toString())
-    }
-}
-
-fun openAlarmApp(context: Context) {
-    try {
-        val intent = Intent(AlarmClock.ACTION_SHOW_ALARMS)
-        context.startActivity(intent)
-    } catch (e: java.lang.Exception) {
-        d("openAlarmApp", e.toString())
-    }
-}
-
-fun openCalendar(context: Context) {
-    try {
-        val cal: Calendar = Calendar.getInstance()
-        cal.time = Date()
-        val time = cal.time.time
-        val builder: Uri.Builder = CalendarContract.CONTENT_URI.buildUpon()
-        builder.appendPath("time")
-        builder.appendPath(time.toString())
-        context.startActivity(Intent(Intent.ACTION_VIEW, builder.build()))
-    } catch (e: Exception) {
-        try {
-            val intent = Intent(Intent.ACTION_MAIN)
-            intent.addCategory(Intent.CATEGORY_APP_CALENDAR)
-            context.startActivity(intent)
-        } catch (e: Exception) {
-            d("openCalendar", e.toString())
-        }
-    }
-}
-
-fun isTablet(context: Context): Boolean {
-    val windowManager = context.getSystemService(Context.WINDOW_SERVICE) as WindowManager
-    val metrics = DisplayMetrics()
-    @Suppress("DEPRECATION")
-    windowManager.defaultDisplay.getMetrics(metrics)
-    val widthInches = metrics.widthPixels / metrics.xdpi
-    val heightInches = metrics.heightPixels / metrics.ydpi
-    val diagonalInches = sqrt(widthInches.toDouble().pow(2.0) + heightInches.toDouble().pow(2.0))
-    if (diagonalInches >= 7.0) return true
-    return false
-}
-
-fun initActionService(context: Context): ActionService? {
-    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
-        val actionService = ActionService.instance()
-        if (actionService != null) {
-            return actionService
-        } else {
-            openAccessibilitySettings(context)
-        }
-    } else {
-        showToastLong(context, "This action requires Android P (9) or higher" )
-    }
-
-    return null
-}
-
-fun openAccessibilitySettings(context: Context) {
-    val intent = Intent(Settings.ACTION_ACCESSIBILITY_SETTINGS)
-    val cs = ComponentName(context.packageName, ActionService::class.java.name).flattenToString()
-    val bundle = Bundle()
-    bundle.putString(":settings:fragment_args_key", cs)
-    intent.apply {
-        addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
-        putExtra(":settings:fragment_args_key", cs)
-        putExtra(":settings:show_fragment_args", bundle)
-    }
-    context.startActivity(intent)
-}
-
-fun showStatusBar(activity: Activity) {
-    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R)
-        activity.window.insetsController?.show(WindowInsets.Type.statusBars())
-    else
-        @Suppress("DEPRECATION", "InlinedApi")
-        activity.window.decorView.apply {
-            systemUiVisibility = View.SYSTEM_UI_FLAG_LAYOUT_STABLE or View.SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN
-        }
-}
-
-fun hideStatusBar(activity: Activity) {
-    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R)
-        activity.window.insetsController?.hide(WindowInsets.Type.statusBars())
-    else {
-        @Suppress("DEPRECATION")
-        activity.window.decorView.apply {
-            systemUiVisibility = View.SYSTEM_UI_FLAG_IMMERSIVE or View.SYSTEM_UI_FLAG_FULLSCREEN
-        }
-    }
-}
-
-/*fun Context.isDarkThemeOn(): Boolean {
-    return resources.configuration.uiMode and
-            Configuration.UI_MODE_NIGHT_MASK == UI_MODE_NIGHT_YES
-}
-
-fun Context.copyToClipboard(text: String) {
-    val clipboardManager = getSystemService(Context.CLIPBOARD_SERVICE) as ClipboardManager
-    val clipData = ClipData.newPlainText(getString(R.string.app_name), text)
-    clipboardManager.setPrimaryClip(clipData)
-    showToastShort(this, "Copied")
-}*/
-
-fun Context.openUrl(url: String) {
-    if (url.isEmpty()) return
-    val intent = Intent(Intent.ACTION_VIEW)
-    intent.data = Uri.parse(url)
-    startActivity(intent)
-}
-
-/*@ColorInt
-fun Context.getColorFromAttr(
-    @AttrRes attrColor: Int,
-    typedValue: TypedValue = TypedValue(),
-    resolveRefs: Boolean = true
-): Int {
-    theme.resolveAttribute(attrColor, typedValue, resolveRefs)
-    return typedValue.data
-}*/
-
-fun uninstallApp(context: Context, appPackage: String) {
-    val intent = Intent(Intent.ACTION_DELETE)
-    intent.data = Uri.parse("package:$appPackage")
-    context.startActivity(intent)
-}
-
-fun dp2px(resources: Resources, dp: Int): Int {
-    return TypedValue.applyDimension(
-        TypedValue.COMPLEX_UNIT_DIP,
-        dp.toFloat(),
-        resources.displayMetrics
-    ).toInt()
-}
-
-fun storeFile(activity: Activity) {
-    val intent = Intent(Intent.ACTION_CREATE_DOCUMENT).apply {
-        addCategory(Intent.CATEGORY_OPENABLE)
-        type = "application/json"
-        putExtra(Intent.EXTRA_TITLE, "backup.json")
-    }
-    ActivityCompat.startActivityForResult(activity, intent, BACKUP_WRITE, null)
-}
-
-fun loadFile(activity: Activity) {
-    val intent = Intent(Intent.ACTION_OPEN_DOCUMENT).apply {
-        addCategory(Intent.CATEGORY_OPENABLE)
-        type = "application/json"
-    }
-    ActivityCompat.startActivityForResult(activity, intent, BACKUP_READ, null)
-}
-
-fun getHexForOpacity(context: Context, prefs: Prefs): Int {
-    var setColor = prefs.opacityNum
-    var isDarkMode = prefs.appTheme.toString()
-
-    var hex = Integer.toHexString(setColor).toString()
-    if (hex.length < 2)
-        hex = "$hex$hex"
-
-    if (isDarkMode.toString() == "System") {
-        return if (isDarkMode(context)) {
-            android.graphics.Color.parseColor("#${hex}000000")
-        } else {
-            android.graphics.Color.parseColor("#${hex}FFFFFF")
-        }
-    }
-    return if (isDarkMode.toString() == "Light") {
-        android.graphics.Color.parseColor("#${hex}FFFFFF")
-    } else {
-        android.graphics.Color.parseColor("#${hex}000000")
-    }
-}
-
-private fun isDarkMode(context: Context): Boolean {
-    val darkModeFlag = context.resources.configuration.uiMode and Configuration.UI_MODE_NIGHT_MASK
-    return darkModeFlag == Configuration.UI_MODE_NIGHT_YES
-}
