--- app/src/main/java/app/mlauncher/style/Color.kt
@@ -1,7 +1,0 @@
-package app.mlauncher.style
-
-import androidx.compose.ui.graphics.Color
-
-val textLight = Color(0xFFFFFFFF)
-val textDark = Color(0xFF000000)
-val textGray = Color(0xFF858585)
