--- app/src/main/java/app/mlauncher/style/Dimen.kt
@@ -1,7 +1,0 @@
-package app.mlauncher.style
-
-import androidx.compose.ui.unit.dp
-
-val CORNER_RADIUS = 2.dp
-val BORDER_SIZE = 1.5.dp
-val SETTINGS_PADDING = 15.dp
