+++ app/src/main/java/com/github/hecodes2much/mlauncher/style/Color.kt
@@ -1,0 +1,7 @@
+package com.github.hecodes2much.mlauncher.style
+
+import androidx.compose.ui.graphics.Color
+
+val textLight = Color(0xFFFFFFFF)
+val textDark = Color(0xFF000000)
+val textGray = Color(0xFF858585)
