+++ app/src/main/java/com/github/hecodes2much/mlauncher/style/Dimen.kt
@@ -1,0 +1,7 @@
+package com.github.hecodes2much.mlauncher.style
+
+import androidx.compose.ui.unit.dp
+
+val CORNER_RADIUS = 2.dp
+val BORDER_SIZE = 1.5.dp
+val SETTINGS_PADDING = 15.dp
