+++ app/src/main/java/com/github/therepoclub/slider/DimensionUtil.kt
@@ -1,0 +1,44 @@
+package com.github.therepoclub.slider
+
+
+/**
+ * [Linear Interpolation](https://en.wikipedia.org/wiki/Linear_interpolation) function that moves
+ * amount from it's current position to start and amount
+ * @param start of interval
+ * @param end of interval
+ * @param amount e closed unit interval [0, 1]
+ */
+fun lerp(start: Float, end: Float, amount: Float): Float {
+    return (1 - amount) * start + amount * (end - start)
+}
+
+/**
+ * Scale x1 from start1..end1 range to start2..end2 range
+ */
+fun scale(start1: Float, end1: Float, pos: Float, start2: Float, end2: Float) =
+    lerp(start2, end2, calculateFraction(start1, end1, pos))
+
+/**
+ * Scale x.start, x.endInclusive from a1..b1 range to a2..b2 range
+ */
+fun scale(
+    start1: Float,
+    end1: Float,
+    range: ClosedFloatingPointRange<Float>,
+    start2: Float,
+    end2: Float
+) =
+    scale(start1, end1, range.start, start2, end2)..scale(
+        start1,
+        end1,
+        range.endInclusive,
+        start2,
+        end2
+    )
+
+
+/**
+ * Calculate fraction for value between a range [end] and [start] coerced into 0f-1f range
+ */
+fun calculateFraction(start: Float, end: Float, pos: Float) =
+    (if (end - start == 0f) 0f else (pos - start) / (end - start)).coerceIn(0f, 1f)
