+++ app/src/main/java/com/github/therepoclub/slider/gesture/MotionEvent.kt
@@ -1,0 +1,5 @@
+package com.github.therepoclub.slider.gesture
+
+enum class MotionEvent {
+    Idle, Down, Move, Up
+}
