+++ app/src/main/java/com/github/therepoclub/slider/ui/Color.kt
@@ -1,0 +1,7 @@
+package com.github.therepoclub.slider.ui
+
+import androidx.compose.ui.graphics.Color
+
+val ThumbColor = Color(0xfff0f0f0)
+val ActiveTrackColor = Color(0xff489cef)
+val InactiveTrackColor = Color(0xffcccccc)
