+++ fastlane/Appfile
@@ -1,2 +1,2 @@
 json_key_file("") # Path to the json secret file - Follow https://docs.fastlane.tools/actions/supply/#setup to get one
-package_name("app.mlauncher") # e.g. com.krausefx.app
+package_name("com.github.hecodes2much.mlauncher") # e.g. com.krausefx.app
